The transition to multi-module has been updated.

+++ kafka-to-elastic/src/main/java/com/cengizhanyavuz/wordvault/kafka/to/elastic/repo/WordElasticRepository.java
@@ -4,5 +4,5 @@ package com.cengizhanyavuz.wordvault.kafka.to.elastic.repo;
 import com.cengizhanyavuz.wordvault.kafka.to.elastic.model.Word;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface WordElasticRepository extends ElasticsearchRepository<Word,Integer> {
+public interface WordElasticRepository extends ElasticsearchRepository<Word,Long> {
 }
